Add assets and navigation menu component (#2)
* ✨ Refactor: Update package name in Android manifest

Updated the package name in the Android manifest file to "com.mushroommentor" to reflect the correct package structure. This change was made to ensure consistency and clarity in the codebase.

* ✨ setup: Add image, and icon assets

* ✨ setup:  Add new colors, sizes, and shadows

ℹ️ Add new color constants, size constants, and shadow configurations for styling purposes.

* ✨ feat: Add navigation menu feature

The NavigationView component was created to display the navigation menu with clickable options. This feature enhances the user experience by providing easy access to different sections of the application.🍄

* ✨ refactor: Add support for importing PNG and JPG files

Added declarations for importing PNG and JPG files in TypeScript by declaring 
modules '*.png' and '*.jpg'. This change was necessary to allow importing 
image files in the project without encountering type errors.