Add menu item type declaration (#4)
* ✨ Add screen for viewing mushroom details

Added a new screen component `MushroomDetails` to display details of specific mushrooms.
Updated the `KnowMushrooms` screen to navigate to `MushroomDetails` on press.

* ✨ Add new screens for Glossary, PestIdentification, NutritionDetails, Recipes & Contact Us

- Added new screens for Glossary, PestIdentification, NutritionDetails,
Recipes, and ContactUs to the application. These screens contain the
basic structure and components needed for each respective screen.

* ✨ Add mushroom data to constants file

Added data for various types of mushrooms along with their name, image URL, and description to the constants file for easier access in the application.

* ✨ Add reusable components for Card, Separator, and Table

* ✨ Update navigation screens and titles

ℹ️ Refactor navigation screens and update titles for better user experience.

* ✨ Add new mushroom gallery URL 🍄

Added new URL for the mushroom gallery in the environment variables file.

* ✨ Remove NavigationView component and related styles

Removed the NavigationView component and its related styles as they are no longer needed and can be refactored for better organization.